delete sql files

+++ src/main/java/info/blockchain/api/Util.java
@@ -1,5 +1,5 @@
 package info.blockchain.api;
 
 public class Util {
-    public static final long SATOSHI_IN_BTC = 100000000;
+    public static final int SATOSHI_IN_BTC = 100000000;
 }
